pyproject.toml
[tool.poetry]
name = "meter-emulator"
version = "0.1.0"
description = "Pluggable energy meter emulator with configurable frontends and backends"
authors = []
packages = [{include = "meter_emulator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.34"}
httpx = "^0.28"
pyyaml = "^6.0"
pydantic = "^2.0"
zeroconf = "^0.146"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.25"
pre-commit = "^4.5.1"
ruff = "^0.15.0"

[tool.poetry.scripts]
meter-emulator = "meter_emulator.main:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
